build: restructure build process
