+++ .eslintignore
@@ -1,2 +1,3 @@
-dist/
+headless-backoffice-bridge.min.js
+lib/
