+++ src/components/index.ts
@@ -1,2 +1,2 @@
-export { default as ImageElement } from './image'
-export { default as TinyMceElement } from './tinymce'
+export { default as ImageElement } from './image/index.js'
+export { default as TinyMceElement } from './tinymce/index.js'
