+++ src/pickers/index.ts
@@ -1,2 +1,2 @@
-export { default as contentPicker, ContentPickerConfig, ContentPickerItem } from './content'
-export { default as mediaPicker, MediaPickerConfig, MediaPickerItem } from './media'
+export { default as contentPicker } from './content/index.js'
+export { default as mediaPicker } from './media/index.js'
