+++ tsconfig.json
@@ -1,16 +1,15 @@
 {
   "compilerOptions": {
-    "target": "es2019",
-    "module": "es2020",
+    "target": "es2020",
+    "module": "esnext",
     "lib": ["es2020", "DOM", "DOM.Iterable"],
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true,
     "inlineSources": true,
     "sourceMap": true,
-    "baseUrl": "./src",
+    "outDir": "./lib",
     "rootDir": "./src",
-    "outDir": "./dist",
     "strict": true,
     "moduleResolution": "node",
     "esModuleInterop": true,
